requirements.txt
asgiref==3.9.1
Django==5.0.14
django-mssql-backend==2.8.1
djangorestframework==3.16.0
djangorestframework_simplejwt==5.5.0
mssql-django==1.5
PyJWT==2.9.0
pyodbc==5.2.0
pytz==2025.2
sqlparse==0.5.3
tzdata==2025.2
